refactor: update mdx helpers
