+++ src/components/content/custom-mdx.tsx
@@ -3,8 +3,7 @@ import * as runtime from 'react/jsx-runtime';
 import { evaluate, type EvaluateOptions } from '@mdx-js/mdx';
 import rehypePrettyCode from 'rehype-pretty-code';
 import remarkGfm from 'remark-gfm';
-
-import { transformers } from '@/lib/highlight-code';
+import type { ShikiTransformer } from 'shiki';
 
 import { mdxComponents } from '@/components/content/mdx-components';
 
@@ -13,6 +12,17 @@ interface CustomMDXProps {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   additionalComponents?: Record<string, React.ComponentType<any>>;
 }
+
+export const transformers = [
+  {
+    code(node) {
+      if (node.tagName === 'code') {
+        const raw = this.source;
+        node.properties['__raw__'] = raw;
+      }
+    },
+  },
+] as ShikiTransformer[];
 
 const rehypePrettyCodeOptions = {
   theme: {
